Add Bash literate config
Prepare for bump to emacs 28.1, updates to mu4e and system keybindings.
I'm rebinding the ctrl key to caps lock to try and avoid emacs pinky.
This config change is at the system level, and so is loaded by way of executing
a bash function. That function is now contained in the bash_functions rc file
and run on startup. The contents of all bash files is now documented in the
Bash.org literate config.