Added a debian manifest and some pretty themes
Spend a /while/ wrestling with dependency problems while attempting to build
emacs 28.1 from source on debian. This commit includes a list of dependencies
that are required to get things working. I dropped a literate config into an
org-roam note, but perhaps what I'll do is to set up a System.org file that
tracks all of the system configuration steps that are being taken when I set up
a new system. This might make life easier in the future.