Editor updates
Dotfile updates for emacs, vim and nvim. Emacs updates are relatively
small, just remove corfu config because autocomplete was causing
headaches when writing.

Add a bunch of cleanup functionality to the vim/nvim configs. Create a
totally seperate, minimal vim config that aims to enhance vim use for
root. Have needed this a few times recently when working on various
servers where a few bits of config might need to be tweaked and we don't
necessarily want to bring the whole enchalada with us. The other vim and
nvim config updates also aim to move in the direction of further
modularization. Starting to write more config in lua and just starting
to scratch the surface of how to write local plugins. Also include a few
updates to the lsp config.

Signed-off-by: Adam McCartney <adam@mur.at>