Add some vcs functionality to zsh
Leverages the builtin vcs info functions to create a prompt that
displays info about the current directory (whether it is in version
control or not, if so, what is the current state of the repo). Note that
this is all happnening synchronously, so it's a little bit sluggish at
the moment.

Also adds some profiling ála https://www.youtube.com/watch?v=cYmKRCUn96I