Tweaks to init.el
Replace the corfu completion backend with company and use set up some
defaults that we can actually live with. The most annoying thing (completion
in ord mode buffers is fixed). Probably need to keep an eye out for automatic
selection, but company seems to not do this either.

Also, get a bit newbie with regards to the "unannoy.el" package. In particular,
it's fine to use the menu bar, so enabling it again.