Tidy up bash rc stuff
This slightly restructures some of the various files that get sourced by bash on
startup. Specifically, it tidies up how git completion and fzf key bindings are
pulled in. Also it fixes the git prompt, reverting to the behaviour where tree
state is displayed dynamically. This was a case of making sure the local vars
used to store the info were reset each time PROMPT_COMMAND was called.