Rework of home services
This adds the following systemd --user services,timers and sockets to guix home:

+ emacs.service
+ mbsync.service + timer
+ muindex.service + timer
+ foot-server.service + socket

Also, we restructured the scripts directories.
+ bin -> mk-scripts
+ scripts -> home-bin/bin

We're also adding home-bin to the home-dotfiles service and syncing our scripts.
On the local machine $HOME/bin was renamed to $HOME/bin-archive incase there are
some scripts that we end up looking for.