Add some fixes/workarounds for bashrc
Slowly figuring out how things should be initialized in the process of initializing
the shell. It's a bit confusing with guix-home as so much stuff (XCURSOR, PS1 for GUIX_ENVIRONMENT)
is set in default files like /etc/profile.d/zzz-guix.sh and the custom bashrc doesn't seem
to get sourced at times. I think the fix is to adapt the home services that set up the profile
and make sure the variables are set as expected there.