Simplify plugin management in neovim
Discovered that vim has build in "packages" functionality. It also provides
":packadd" and ":packadd!" commands for loading plugins from directories
listed on the "packpath".

Useful information from Wincent here:
https://www.youtube.com/watch?v=X2_R3uxDN6g

So, this clears all plugins out of the vimrc and replaces (some) of them with
git submodules. This makes it a bit more controlled in terms of managing the
versions. It also plays nice(r) with guix home.