feature/luasnip integration
Begin trying to figure out how this luasnip thing works and see if
it can be useful for the workflow. So far so head-scratchy, did some
initial testing today but was left underwhelmed. It's a slightly differnt
kettle of fish to yassnippet and the custom snippets we had been using for
vim.

Cleaned up config of nvim-cmp (i.e. copy-pasta'd a bunch more stuff from
wincent's dotfile).