Use new defaults from nvim 0.11.0 Nvim 0.11.0 has a couple of cool improvements related to lsp and completion. The story is that some popular plugins have been adopted into the neovim core repo and are now available by default.
For instance, the plugin lspconfig is no longer needed to specify lsp configurations.
Similar functionality is provided by the "vim.lsp" api.

Similar changes have happened in the world of completion. It's possible to turn on
some pretty useful completion with the default setup.

See this blog post for an overview of what's new (or run :h news)
https://gpanders.com/blog/whats-new-in-neovim-0-11/#builtin-auto-completion